Moved exponential search logic into its own file
